+++ R/metacor.R
@@ -61,8 +61,9 @@
 #' @param null.effect A numeric value specifying the effect under the
 #'   null hypothesis.
 #' @param method.bias A character string indicating which test is to
-#'   be used.  Either \code{"rank"}, \code{"linreg"}, or \code{"mm"},
-#'   can be abbreviated.  See function \code{\link{metabias}}
+#'   be used. Either \code{"Begg"}, \code{"Egger"}, or
+#'   \code{"Thompson"}, can be abbreviated. See function
+#'   \code{\link{metabias}}.
 #' @param backtransf A logical indicating whether results for Fisher's
 #'   z transformed correlations (\code{sm = "ZCOR"}) should be back
 #'   transformed in printouts and plots. If TRUE (default), results
@@ -553,7 +554,7 @@ metacor <- function(cor, n, studlab,
   ##
   chknumeric(null.effect, length = 1)
   ##
-  method.bias <- setchar(method.bias, .settings$meth4bias)
+  method.bias <- setmethodbias(method.bias)
   ##
   chklogical(backtransf)
   ##