+++ R/metamean.R
@@ -83,8 +83,9 @@
 #' @param tau.common A logical indicating whether tau-squared should
 #'   be the same across subgroups.
 #' @param method.bias A character string indicating which test is to
-#'   be used.  Either \code{"rank"}, \code{"linreg"}, or \code{"mm"},
-#'   can be abbreviated.  See function \code{\link{metabias}}
+#'   be used. Either \code{"Begg"}, \code{"Egger"}, or
+#'   \code{"Thompson"}, can be abbreviated. See function
+#'   \code{\link{metabias}}.
 #' @param backtransf A logical indicating whether results should be
 #'   back transformed in printouts and plots for \code{sm = "MLN"}. If
 #'   TRUE (default), results will be presented as means; otherwise
@@ -671,7 +672,7 @@ metamean <- function(n, mean, sd, studlab,
   ##
   chknumeric(null.effect, length = 1)
   ##
-  method.bias <- setchar(method.bias, .settings$meth4bias)
+  method.bias <- setmethodbias(method.bias)
   ##
   if (!is.null(text.fixed))
     chkchar(text.fixed, length = 1)