+++ R/update.meta.R
@@ -92,10 +92,8 @@
 #' @param null.effect A numeric value specifying the effect under the
 #'   null hypothesis.
 #' @param method.bias A character string indicating which test for
-#'   funnel plot asymmetry is to be used. Either \code{"rank"},
-#'   \code{"linreg"}, \code{"mm"}, \code{"count"}, \code{"score"}, or
-#'   \code{"peters"}, can be abbreviated. See function
-#'   \code{\link{metabias}}
+#'   funnel plot asymmetry is to be used, can be abbreviated. See
+#'   function \code{\link{metabias}.}
 #' @param backtransf A logical indicating whether results should be
 #'   back transformed in printouts and plots. If \code{backtransf =
 #'   TRUE}, results for \code{sm = "OR"} are printed as odds ratios